SCC: Move `find_driver_loops` to `transformations.utilities`
